Reenable linter settings and fix PR comments
